Update source_extraction.h
unsigned long int num_coords

+++ source_extraction.h
@@ -31,10 +31,10 @@ extern "C" {
 void source_extraction(int rank, unsigned int facet, likelihood_params *par, complexd *facet_vis, double *facet_sigma2, 
                        double *sum_w, double l0, double m0, double flux, double mu, double e1, double e2,                        
                        complexd *visSkyMod, complexd *visData, complexd *visGal, float *sigma2_vis, bool *flag,  
-                       unsigned int num_coords, double *uu_metres, double *vv_metres, double *ww_metres, double len);
+                       unsigned long int num_coords, double *uu_metres, double *vv_metres, double *ww_metres, double len);
 #else
 void source_extraction(double l0, double m0, double flux, double mu, double e1, double e2, likelihood_params *par, complexd *visSkyMod, 
-                       complexd *visData, complexd *visGal, float *sigma2_vis, unsigned int num_coords, 
+                       complexd *visData, complexd *visGal, float *sigma2_vis, unsigned long int num_coords, 
                        double *uu_metres, double *vv_metres, double *ww_metres);
 #endif
